refactor: 사용자 서류평가 쪽 ssr 적용 (#71)
