+++ apps/web/src/app/(main)/apply-management/[tab]/[id]/_components/ApplicantDetail/ApplicantDetail.css.ts
@@ -6,7 +6,7 @@ export const container = style({
   padding: '4.2rem 3.2rem',
   display: 'flex',
   flexDirection: 'column',
-  gap: '10rem',
+  gap: '2.4rem',
   justifyContent: 'center',
   alignItems: 'flex-start',
   borderRadius: '24px',
