+++ apps/web/src/app/(main)/docs-evaluation/application/[id]/page.tsx
@@ -1,155 +1,27 @@
-'use client';
-import { useParams, useRouter, useSearchParams } from 'next/navigation';
-import { useEffect, useRef, useState } from 'react';
-import ApplicantDetail from '@web/app/(main)/apply-management/[tab]/[id]/_components/ApplicantDetail/ApplicantDetail';
+// app/docs-evaluation/application/[id]/page.tsx
+import React from 'react';
+import { ServerFetchBoundary } from '@web/store/query/ServerFetchBoundary';
+import { getApplicationDetailQueryOptions } from '@web/store/query/useApplicationDetailQuery';
+import { getServerSideTokens } from '@web/api/serverSideTokens';
+import DetailClient from './DetailClient';
 
-import { Flex } from '@repo/ui/Flex';
-import { DocsDetailHeader } from '@web/app/(main)/docs-evaluation/application/[id]/_components/DocsDetailHeader/DocsDetailHeader';
-import {
-  DocsEvaluation,
-  EvaluationData,
-} from '@web/app/(main)/docs-evaluation/application/[id]/_components/DocsEvaluation/DocsEvaluation';
-import * as styles from './page.css';
-import { EvaluationAddCommentCard } from './_components/EvaluationAddCommentCard/EvaluationAddCommentCard';
-import {
-  Evaluation,
-  useApplicationDetailQuery,
-} from '@web/store/query/useApplicationDetailQuery';
-import { useBulkEvaluationsMutation } from '@web/store/mutation/useBulkEvaluations';
-import { useUserStore } from '@web/store/state/userStore';
-import { useToggleAcquaintanceMutation } from '@web/store/mutation/useToggleAcquaintanceMutation';
+interface PageProps {
+  params: { id: string };
+}
 
-export default function Page() {
-  const router = useRouter();
-  const params = useParams();
+export default async function Page({ params }: PageProps) {
   const applicationId = Number(params.id);
+  const tokens = await getServerSideTokens();
 
-  const {
-    data: application,
-    isLoading,
-    isError,
-  } = useApplicationDetailQuery(applicationId);
-  const name = useUserStore.getState().name;
-  const myUserId = useUserStore.getState().userId;
-  console.log('사용자', myUserId);
-  const [isRelation, setIsRelation] = useState(false);
-
-  // 3) 서버에서 데이터 로드 완료되면 초기화
-  useEffect(() => {
-    if (application) {
-      const rel = application.acquaintances.some((a) => a.userId === myUserId);
-      setIsRelation(rel);
-    }
-  }, [application, myUserId]);
-
-  const toggleAcq = useToggleAcquaintanceMutation(applicationId);
-
-  const handleToggle = () => {
-    toggleAcq.mutate(undefined, {
-      onSuccess: (acquainted: boolean) => {
-        console.log('API 응답 acquainted:', acquainted);
-        setIsRelation(acquainted);
-      },
-      onError: (err) => {
-        console.error('토글 실패:', err);
-        // 필요하면 롤백하거나 에러 메시지 띄우기
-      },
-    });
-  };
-
-  const [scores, setScores] = useState<number[]>([]);
-
-  // 3) Bulk Mutation
-  const mutation = useBulkEvaluationsMutation(applicationId);
-
-  const average = application?.documentAverageScore;
-  const criteriaList = application?.documentEvaluationCriterias ?? [];
-  const evaluationList: Evaluation[] = criteriaList.map((c) => ({
-    id: c.id,
-    score: c.score ?? 5,
-    criteria: {
-      ...c,
-      type: c.type as 'DOCUMENT' | 'INTERVIEW',
-    },
-    user: { userId: 0, name: '', profileColor: '' },
-  }));
-
-  const myComments = application?.documentComments.filter(
-    (c) => c.user.userId === myUserId
-  );
-
-  useEffect(() => {
-    if (evaluationList.length > 0) {
-      setScores(evaluationList.map((e) => e.score ?? 5));
-    }
-  }, [evaluationList.length]);
-
-  if (isLoading) return null;
-  if (isError || !application) {
-    router.push('/404');
-    return null;
-  }
-
-  const handleSave = () => {
-    // 1) 보낼 페이로드 생성
-    const payload = {
-      applicationId,
-      evaluations: evaluationList.map((e, idx) => ({
-        criteriaId: e.criteria.id,
-        score: scores[idx] ?? 0,
-      })),
-    };
-
-    // 2) 콘솔에 찍기
-    console.log('[DocsEvaluation] handleSave payload:', payload);
-
-    // 3) 실제 호출
-    mutation.mutate(payload);
-  };
-
-  //console.log('지원서 디테일', application);
-
-  const handleScoreChange = (name: string, next: number) => {
-    const idx = Number(name.split('-')[1]);
-    setScores((prev) => {
-      const copy = [...prev];
-      copy[idx] = next;
-      return copy;
-    });
-  };
-
-  // 6) DocsEvaluation에 넘길 평가 리스트 포맷
-  // DocsEvaluation에 넘길 데이터
-  const documentEvaluationData: EvaluationData = {
-    evaluationType: 'score',
-    evaluationList,
-  };
+  const detailOptions = getApplicationDetailQueryOptions({
+    applicationId,
+    tokens,
+  });
 
   return (
-    <div className={styles.container}>
-      {/* 헤더 */}
-      <DocsDetailHeader
-        isChecked={isRelation}
-        onToggle={handleToggle}
-        name={application.name}
-      />
-
-      <Flex gap="2rem">
-        <ApplicantDetail application={application} />
-
-        <div className={styles.rightSection}>
-          <EvaluationAddCommentCard comments={myComments!} />
-        </div>
-      </Flex>
-
-      {/* 문서 평가 스코어링 */}
-      <DocsEvaluation
-        average={average!}
-        evaluationData={documentEvaluationData}
-        scores={scores}
-        onScoreChange={handleScoreChange}
-        onSave={handleSave}
-      />
-    </div>
+    <ServerFetchBoundary fetchOptions={detailOptions}>
+      {/* 이 컴포넌트는 use client; 로 선언 */}
+      <DetailClient />
+    </ServerFetchBoundary>
   );
 }
