+++ apps/web/src/store/query/useApplicationsQuery.ts
@@ -1,14 +1,13 @@
 import {
-  keepPreviousData,
+  UseQueryResult,
   queryOptions,
+  useQuery,
   useSuspenseQuery,
-  type UseSuspenseQueryOptions,
 } from '@tanstack/react-query';
 import { GET } from '@web/api/fetch';
-import type { ApiResponse } from '@web/api/types';
+import type { ApiResponse, Tokens } from '@web/api/types';
 import { queryKeys } from '../constants';
 
-/** 페이징 메타 정보 */
 export interface PaginationMeta {
   page: number;
   size: number;
@@ -17,7 +16,6 @@ export interface PaginationMeta {
   isLast: boolean;
 }
 
-/** 사용자용 지원서 요약 DTO */
 export interface ApplicationSummary {
   documentResultAnnounced: boolean;
   id: number;
@@ -35,13 +33,11 @@ export interface ApplicationSummary {
   interviewSchedule?: string;
 }
 
-/** API 결과 - 페이징 응답 */
 export interface ApplicationsResult {
   data: ApplicationSummary[];
   pagination: PaginationMeta;
 }
 
-/** 전체 API 응답 래퍼 */
 export interface ApplicationsResponse extends ApiResponse<ApplicationsResult> {}
 
 /** useApplicationsQuery 옵션 */
@@ -49,26 +45,31 @@ export interface UseApplicationsQueryOptions {
   recruitmentId: number;
   evaluationStatus?: 'ALL' | 'EVALUATED' | 'NOT_EVALUATED';
   keyword?: string;
-  page?: number; // 0-based
+  page?: number;
   size?: number;
+  tokens?: Tokens;
 }
 
-const APPS_STALE_TIME = 1000 * 60 * 3; // 3분
+const APPS_STALE_TIME = 1000 * 60 * 1;
+const APPS_CACHE_TIME = 1000 * 60 * 2;
 
 /**
  * 사용자용 지원서 목록 조회 옵션 생성기
  */
-export function getApplicationsQueryOptions({
-  recruitmentId,
-  evaluationStatus = 'ALL',
-  keyword = '',
-  page = 0,
-  size = 9,
-}: UseApplicationsQueryOptions): UseSuspenseQueryOptions<
-  ApplicationsResult,
-  Error
-> {
-  return queryOptions<ApplicationsResult, Error>({
+export function getApplicationsQueryOptions(
+  params: UseApplicationsQueryOptions
+) {
+  const {
+    recruitmentId,
+    evaluationStatus = 'ALL',
+    keyword = '',
+    page = 0,
+    size = 9,
+    tokens,
+  } = params;
+  const pageParam = page + 1;
+
+  return {
     queryKey: queryKeys.applications.userList(
       recruitmentId,
       evaluationStatus,
@@ -82,17 +83,29 @@ export function getApplicationsQueryOptions({
         {
           evaluationStatus,
           keyword,
-          page: String(page + 1),
+          page: String(pageParam),
           size: String(size),
-        }
+        },
+        tokens
       ).then((res) => res.result),
     staleTime: APPS_STALE_TIME,
-    placeholderData: keepPreviousData,
+    gcTime: APPS_CACHE_TIME,
+    keepPreviousData: true,
     refetchOnMount: true,
     enabled: recruitmentId > 0,
-  });
+  };
 }
 
 export function useApplicationsQuery(params: UseApplicationsQueryOptions) {
-  return useSuspenseQuery(getApplicationsQueryOptions(params));
+  return useSuspenseQuery<ApplicationsResult, Error>(
+    getApplicationsQueryOptions(params)
+  );
+}
+
+export function useApplicationsClientQuery(
+  params: UseApplicationsQueryOptions
+): UseQueryResult<ApplicationsResult, Error> {
+  return useQuery<ApplicationsResult, Error>(
+    getApplicationsQueryOptions(params)
+  );
 }
